Renamed benchmark_problems.py -> prblems.py, removed standalone uvw benchmark file.
